Begin splitting project into multiple packages
