+++ package.json
@@ -1,150 +1,28 @@
 {
-  "name": "react-router-hibernate",
-  "version": "0.0.2",
-  "description": "A react-router Switch which can leave inactive routes mounted-but-inactive until you navigate back",
-  "keywords": [],
+  "name": "react-hibernate-workspace",
+  "private": true,
   "license": "MIT",
-  "homepage": "https://github.com/spautz/react-router-hibernate#readme",
+  "homepage": "https://github.com/spautz/react-hibernate#readme",
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/spautz/react-router-hibernate.git"
+    "url": "git+https://github.com/spautz/react-hibernate.git"
   },
   "bugs": {
-    "url": "https://github.com/spautz/react-router-hibernate/issues"
+    "url": "https://github.com/spautz/react-hibernate/issues"
   },
   "author": "Steven Pautz <spautz@gmail.com>",
-  "files": [
-    "dist/",
-    "src/",
-    "LICENSE",
-    "README.md"
+  "workspaces": [
+    "packages/*"
   ],
-  "source": "src/index.ts",
-  "main": "dist/index.cjs.js",
-  "module": "dist/index.esm.js",
-  "jsnext:main": "dist/index.esm.js",
-  "types": "dist/src/index.d.ts",
   "sideEffects": false,
   "scripts": {
-    "____ BASE COMMANDS _________________________________________________": "",
-    "build": "microbundle build --jsx React.createElement",
-    "build:watch": "microbundle watch --jsx React.createElement",
-    "clean": "rimraf dist/ storybook-static/",
-    "format": "prettier --write \"**/*.*\"",
-    "format:check": "prettier --list-different \"**/*.*\"",
-    "lint": "eslint \"**/*.{js,jsx,json,ts,tsx}\"",
-    "prepare": "yon build",
-    "release": "standard-version --sign --release-as ",
-    "storybook": "start-storybook -p 6006",
-    "build-storybook": "build-storybook",
-    "test": "jest",
-    "test:clean": "rimraf coverage/",
-    "test:coverage": "jest --coverage",
-    "test:report": "jest --coverage --coverageReporters=text-lcov | coveralls",
-    "test:watch": "jest --watch",
-    "types": "tsc --noEmit --p tsconfig.json --jsx react",
-    "____ HOOKS _________________________________________________________": "",
-    "prebuild": "yon clean",
-    "prerelease": "yon clean",
-    "prepublishOnly": "yon check && yon build",
-    "pretest": "yon test:clean",
-    "____ INTEGRATION ___________________________________________________": "",
-    "dev": "yon run format && yon run types && yon run lint",
-    "check": "yon format:check && yon run types && yon run lint",
-    "ci": "yon run check && yon run test:report && yon run build && yon run build-storybook",
-    "all": "yon run dev && yon run test:coverage && yon run build && yon run build-storybook"
+    "prepare": "lerna bootstrap"
   },
-  "dependencies": {
-    "limited-cache": "^0.3.0",
-    "prop-types": "^15.7.2",
-    "react-is": "^16.12.0",
-    "react-reverse-portal": "^1.0.5"
-  },
+  "dependencies": {},
   "devDependencies": {
-    "@babel/core": "^7.8.4",
-    "@material-ui/core": "^4.9.4",
-    "@storybook/addon-knobs": "^5.3.14",
-    "@storybook/addons": "^5.3.14",
-    "@storybook/preset-typescript": "^1.2.0",
-    "@storybook/react": "^5.3.14",
-    "@testing-library/jest-dom": "^5.1.1",
-    "@testing-library/react": "^9.4.1",
-    "@types/history": "^4.7.5",
-    "@types/jest": "^25.1.3",
-    "@types/node": "^13.7.6",
-    "@types/react": "^16.9.23",
-    "@types/react-is": "^16.7.1",
-    "@types/react-redux": "^7.1.7",
-    "@types/react-router": "^5.1.4",
-    "@types/react-router-dom": "^5.1.3",
-    "@typescript-eslint/eslint-plugin": "^2.21.0",
-    "@typescript-eslint/parser": "^2.21.0",
-    "babel-loader": "^8.0.6",
-    "coveralls": "^3.0.9",
-    "eslint": "^6.8.0",
-    "eslint-config-prettier": "^6.10.0",
-    "eslint-plugin-prettier": "^3.1.2",
-    "eslint-plugin-react": "^7.18.3",
-    "fork-ts-checker-webpack-plugin": "^4.0.4",
-    "history": "^4.10.1",
-    "husky": "^4.2.3",
-    "jest": "^25.1.0",
-    "lint-staged": "^10.0.7",
-    "microbundle": "^0.12.0-next.8",
-    "prettier": "^1.19.1",
-    "react": "^16.12.0",
-    "react-docgen-typescript-loader": "^3.6.0",
-    "react-dom": "^16.12.0",
-    "react-redux": "^7.2.0",
-    "react-router-dom": "^5.1.2",
-    "react-test-renderer": "^16.12.0",
-    "redux": "^4.0.5",
-    "redux-devtools-extension": "^2.13.8",
-    "rimraf": "^3.0.2",
-    "standard-version": "^7.1.0",
-    "ts-jest": "^25.2.1",
-    "ts-loader": "^6.2.1",
-    "tslib": "^1.10.0",
-    "typeface-roboto": "^0.0.75",
-    "typescript": "^3.7.5",
-    "yarn-or-npm": "^3.0.1"
+    "lerna": "^3.20.2"
   },
-  "peerDependencies": {
-    "react": ">=16.8.0",
-    "react-dom": ">=16.8.0",
-    "react-redux": ">=7.0.0",
-    "react-router": ">=5.0.0"
-  },
-  "peerDependenciesMeta": {
-    "react-redux": {
-      "optional": true
-    }
-  },
-  "jest": {
-    "transform": {
-      "^.+\\.tsx?$": "ts-jest"
-    },
-    "testRegex": "(/__tests__/.*|/tests/.*|(\\.|/)(test|spec))\\.[jt]sx?$",
-    "moduleFileExtensions": [
-      "ts",
-      "tsx",
-      "js",
-      "jsx",
-      "json",
-      "node"
-    ],
-    "modulePathIgnorePatterns": [
-      "dist/"
-    ],
-    "collectCoverage": true,
-    "coveragePathIgnorePatterns": [
-      "/util/"
-    ],
-    "coverageReporters": [
-      "json",
-      "html"
-    ]
-  },
+  "peerDependencies": {},
   "lint-staged": {
     "**/*.{css,html,js,jsx,json,less,md,scss,ts,tsx}": [
       "prettier --write"
