+++ packages/react-router-hibernate/src/HibernatingSwitch.tsx
