+++ packages/react-router-hibernate/src/index.ts
@@ -4,4 +4,4 @@ export { default as HibernatingRoute } from './HibernatingRoute';
 export * from './HibernatingRoute';
 
 // @TODO: This could probably be a separate package
-export * from './StaticContainers';
+export * from '../../react-static-containers/src';
