+++ packages/react-router-hibernate/src/renderRoute.ts
