+++ packages/react-router-hibernate/tests/componentCompatibility.tsx
@@ -12,7 +12,7 @@ import {
   restoreReactUnmountErrors,
   suppressReactUnmountErrors,
   renderWithRouterContext,
-} from '../helpers/test';
+} from '../../../helpers/test';
 
 describe('Compatibility with react-router components', () => {
   // We only suppress the unmount errors between tests: if they happen within any test, we want to know
