+++ packages/react-static-containers/src/StaticComponentContainer.tsx
