+++ packages/react-static-containers/src/StaticReduxContainer.tsx
