+++ packages/react-static-containers/src/index.ts
