+++ packages/react-static-containers/src/types.ts
